Event processors restructuration
