+++ packages/geoview-core/src/api/event-processors/abstract-event-processor.ts
