+++ packages/geoview-core/src/api/event-processors/event-processor-children/app-event-processor.ts
@@ -1,5 +1,5 @@
 import { GeoViewStoreType } from '@/core/stores/geoview-store';
-import { AbstractEventProcessor } from './abstract-event-processor';
+import { AbstractEventProcessor } from '../abstract-event-processor';
 import { getGeoViewStore } from '@/core/stores/stores-managers';
 import { NotificationDetailsType } from '@/core/types/cgpv-types';
