+++ packages/geoview-core/src/core/stores/store-interface-and-intial-values/ui-state.ts
@@ -35,7 +35,7 @@ export function initializeUIState(set: TypeSetStore, get: TypeGetStore): IUIStat
         });
       },
     },
-  };
+  } as IUIState;
 
   return init;
 }
