+++ packages/geoview-core/src/core/stores/stores-managers.ts
@@ -3,7 +3,7 @@ import { useContext } from 'react';
 import { create, createStore } from 'zustand';
 import { mountStoreDevtool } from 'simple-zustand-devtools';
 
-import { initializeEventProcessors } from '@/api/eventProcessors';
+import { initializeEventProcessors } from '@/api/event-processors';
 import { TypeMapFeaturesConfig } from '../types/global-types';
 import { IGeoViewState, GeoViewStoreType, geoViewStoreDefinitionWithSubscribeSelector } from './geoview-store';
 import { MapContext } from '@/core/app-start';
